[frontend] Detach button from sidebar and add processing when users is empty
